requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.3
djangorestframework==3.16.0
idna==3.10
importlib_resources==6.5.2
pillow==11.2.1
python-barcode==0.15.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.2
qrcode==8.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
